+++ demos/se051h_nfc_comm_prov/cmake_build/main.c
@@ -91,10 +91,16 @@ static void print_help() {
   printf(" --delete_key <HEX_KEYID>     ==> Delete key with specified hex key "
          "ID (e.g., 0x7FFF3002). \n");
   printf(" --do_readidlist              ==> Read ID list from SE05x. \n");
-  printf(" --t4t_enable_read            ==> Enable Contact less Read. (Not to be used with PCSC interface) \n");
-  printf(" --t4t_disable_read           ==> Disable Contact less Read. (Not to be used with PCSC interface) \n");
-  printf(" --t4t_enable_write           ==> Enable Contact less Write. (Not to be used with PCSC interface) \n");
-  printf(" --t4t_disable_write          ==> Disable Contact less Write. (Not to be used with PCSC interface) \n");
+  printf(" --t4t_enable_read            ==> Enable Contact less Read. (Not to "
+         "be used with PCSC interface) \n");
+  printf(" --t4t_disable_read           ==> Disable Contact less Read. (Not to "
+         "be used with PCSC interface) \n");
+  printf(" --t4t_enable_write           ==> Enable Contact less Write. (Not to "
+         "be used with PCSC interface) \n");
+  printf(" --t4t_disable_write          ==> Disable Contact less Write. (Not "
+         "to be used with PCSC interface) \n");
+  printf(" --doreset-cryptoobjects      ==> Delete all crypto objects. (Any "
+         "other inputs to example will be ignored.) \n");
 
   return;
 }
@@ -237,6 +243,7 @@ sss_status_t ex_sss_entry(ex_sss_boot_ctx_t *pCtx) {
   uint8_t provision_verifiers = 0;
   int parameter_error = 1;
   uint8_t se05x_t4t_access_ctrl_option = se05x_t4t_ac_invalid;
+  uint8_t doresetcryproobjects = 0;
 
   uint8_t dac_key[256] = {0};
   uint8_t *dac_key_ptr = NULL;
@@ -439,6 +446,9 @@ sss_status_t ex_sss_entry(ex_sss_boot_ctx_t *pCtx) {
       } else if (strcmp(argv[i], "--t4t_disable_write") == 0) {
         se05x_t4t_access_ctrl_option = se05x_t4t_ac_disable_write;
         parameter_error = 0;
+      } else if (strcmp(argv[i], "--doreset-cryptoobjects") == 0) {
+        doresetcryproobjects = 1;
+        parameter_error = 0;
       } else {
         parameter_error = 1;
         break;
@@ -456,7 +466,8 @@ sss_status_t ex_sss_entry(ex_sss_boot_ctx_t *pCtx) {
   if (do_reset == 0 && do_ec_key_provision == 0 && do_user_id_provision == 0 &&
       do_aes_key_provision == 0 && only_t4t_provision == 0 &&
       do_delete_key == 0 && do_readidlist == 0 &&
-      se05x_t4t_access_ctrl_option == 0) {
+      se05x_t4t_access_ctrl_option == 0 &&
+      doresetcryproobjects == 0) {
     if (device_network_type == invalidNetworkInterface) {
       printf(
           "Specify at-least one network interface type (--wifi_net_interface "
@@ -472,7 +483,7 @@ sss_status_t ex_sss_entry(ex_sss_boot_ctx_t *pCtx) {
       do_ec_key_provision, do_aes_key_provision, do_user_id_provision,
       provision_with_policy, dac_key_ptr, dac_key_len, dac_cert_ptr,
       dac_cert_len, provision_verifiers, do_delete_key, delete_keyid,
-      do_readidlist, se05x_t4t_access_ctrl_option);
+      do_readidlist, se05x_t4t_access_ctrl_option, doresetcryproobjects);
 
   return kStatus_SSS_Success;
 }